refactor: port optedout check to onboarding service
I've moved the logic to check if a user is opted out to onboarding
(which is its home).

We should not have a monolith service, that is one service should not
have access to another service's database/table.

The distributed monolith has been broken by calling an internal endpoint
exposed by onboarding to get the opted out data of a phone number.

Signed-off-by: Dumbledore <mathenge@healthcloud.co.ke>