feat: add Be.Well app identifier to OTP verification messages
I've added the first 11 hash characters of our Be.Well app to the messages sent for phone number verification.

With this in place, the app can put in place an API to read the app's
identifier and automatically fill in the OTP verification code sent. The
user does
not have to input the code manually as before. For more information
read: https://developers.google.com/identity/sms-retriever/request

Signed-off-by: Dumbledore <mathenge@healthcloud.co.ke>